+++ src/pydbm/models/base.py
@@ -14,15 +14,14 @@ class DbmModel(metaclass=Meta):
     if typing.TYPE_CHECKING:
         required_fields: typing.ClassVar[list[str]]
         objects: typing.ClassVar[DatabaseManager]
-        pk: typing.ClassVar[str]
         id: str
         empty_model: typing.ClassVar[bool]
 
     def __init__(self, **fields: typing.Any) -> None:
-        self.id = fields.pop("pk")
-        self.fields: dict[str, typing.Any] = fields
+        self.id = fields.pop("id")
+        self.fields: dict[str, typing.Any] = fields.copy()
 
-        for key, value in fields.items():
+        for key, value in self.fields.items():
             setattr(self, key, value)
 
     def __repr__(self):
@@ -31,28 +30,28 @@ class DbmModel(metaclass=Meta):
 
     def __eq__(self, other):
         if isinstance(other, type(self)):
-            return self.fields == other.fields and self.pk == other.pk
+            return self.fields == other.fields and self.id == other.id
         return False
 
     def __hash__(self):
-        if self.pk is None:
+        if self.id is None:
             raise TypeError("Model instances without primary key value are unhashable")
-        return hash(self.pk)
+        return hash(self.id)
 
     def __len__(self):
         return self.objects.count()  # type: ignore
 
     def save(self) -> None:
-        self.objects.save(pk=self.pk, fields=self.fields)
+        self.objects.save(id=self.id, fields=self.fields)
 
     def update(self, **updated_fields) -> None:
         for field_name, field_value in updated_fields.items():
             setattr(self, field_name, field_value)
 
-        self.objects.update(pk=self.pk, **updated_fields)
+        self.objects.update(id=self.id, **updated_fields)
 
     def delete(self) -> None:
-        self.objects.delete(pk=self.pk)
+        self.objects.delete(id=self.id)
 
     def as_dict(self) -> dict[str, typing.Any]:
         return self.fields